Split lots of UI pieces out to their own components
