+++ frontend/src/components/BillCard/style.css
@@ -1,0 +1,188 @@
+/* Bill Card styles - extracted from main.css */
+.bill-card {
+  background: var(--card-bg);
+  padding: 1.5rem;
+  border-radius: 8px;
+  box-shadow: var(--shadow);
+  position: relative;
+  transition: transform 0.1s ease-out, background-color 0.3s;
+  will-change: box-shadow;
+}
+
+.bill-card:hover {
+  transform: translateY(-2px);
+}
+
+/* Bill status colors */
+.bill-card.due-today {
+  background: linear-gradient(135deg, #fff9e6 0%, var(--card-bg) 100%);
+  border-left: 4px solid #f39c12;
+}
+
+.bill-card.overdue {
+  background: linear-gradient(135deg, #ffe6e6 0%, var(--card-bg) 100%);
+  border-left: 4px solid #e74c3c;
+}
+
+.bill-card.overdue .amount {
+  color: #e74c3c;
+}
+
+.bill-card.due-today .amount {
+  color: #f39c12;
+}
+
+.bill-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  margin-bottom: 0.5rem;
+}
+
+.bill-header h3 {
+  margin: 0;
+  flex: 1;
+}
+
+.bill-actions {
+  display: flex;
+  gap: 0.5rem;
+  opacity: 0;
+  transition: opacity 0.2s;
+}
+
+.bill-card:hover .bill-actions {
+  opacity: 1;
+}
+
+.action-btn {
+  background: none;
+  border: none;
+  cursor: pointer;
+  font-size: 1.2rem;
+  padding: 0.25rem;
+  border-radius: 4px;
+  transition: all 0.2s;
+  line-height: 1;
+}
+
+.action-btn:hover {
+  background: var(--bg-color);
+  transform: scale(1.1);
+}
+
+.edit-btn:hover {
+  background: rgba(74, 144, 226, 0.1);
+}
+
+.delete-btn:hover {
+  background: rgba(231, 76, 60, 0.1);
+}
+
+.bill-card .amount {
+  font-size: 1.5rem;
+  font-weight: bold;
+  color: var(--primary-color);
+  margin: 0.5rem 0;
+}
+
+.bill-card .due-date {
+  color: #666;
+  font-size: 0.9rem;
+}
+
+.bill-card .notes {
+  color: #888;
+  font-size: 0.85rem;
+  font-style: italic;
+  margin-top: 0.5rem;
+  line-height: 1.4;
+}
+
+.bill-footer {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  margin-top: 1rem;
+  flex-wrap: wrap;
+}
+
+.btn-mark-paid {
+  padding: 0.5rem 1rem;
+  border: 2px solid var(--success-color);
+  background: white;
+  color: var(--success-color);
+  border-radius: 6px;
+  font-size: 0.85rem;
+  font-weight: 600;
+  cursor: pointer;
+  transition: all 0.2s;
+  margin-left: auto;
+}
+
+.btn-mark-paid:hover {
+  background: var(--success-color);
+  color: white;
+  transform: translateY(-1px);
+}
+
+.btn-mark-paid.paid {
+  background: var(--success-color);
+  color: white;
+  border-color: var(--success-color);
+}
+
+.btn-mark-paid.paid:hover {
+  background: white;
+  color: var(--success-color);
+}
+
+.status {
+  padding: 0.25rem 0.75rem;
+  border-radius: 20px;
+  font-size: 0.8rem;
+  font-weight: 600;
+  display: inline-block;
+}
+
+.status.paid {
+  background: #d4edda;
+  color: #155724;
+}
+
+.status.unpaid {
+  background: #f8d7da;
+  color: #721c24;
+}
+
+.bill-card .category {
+  display: inline-block;
+  padding: 0.25rem 0.75rem;
+  background: var(--bg-color);
+  border-radius: 20px;
+  font-size: 0.85rem;
+  color: #666;
+  margin-top: 0.5rem;
+}
+
+.bill-card .badge {
+  display: inline-block;
+  padding: 0.25rem 0.5rem;
+  background: var(--secondary-color);
+  color: white;
+  border-radius: 4px;
+  font-size: 0.75rem;
+  font-weight: 600;
+  margin-top: 0.5rem;
+  margin-left: 0.5rem;
+}
+
+/* View button in bills list */
+.view-btn {
+  background: rgba(74, 144, 226, 0.1);
+}
+
+.view-btn:hover {
+  background: rgba(74, 144, 226, 0.2);
+  color: var(--primary-color);
+}
