+++ frontend/src/components/EmptyState/index.jsx
@@ -1,0 +1,9 @@
+import "./style.css";
+
+export default function EmptyState({ message, children }) {
+  return (
+    <div class="empty-state">
+      {children || <p>{message}</p>}
+    </div>
+  );
+}
