+++ frontend/src/components/EmptyState/style.css
@@ -1,0 +1,6 @@
+/* Empty State styles - extracted from main.css */
+.empty-state {
+  text-align: center;
+  padding: 4rem 2rem;
+  color: #666;
+}
