+++ frontend/src/components/PaymentActionsModal/style.css
@@ -1,0 +1,14 @@
+/* Payment Actions */
+.payment-actions-buttons {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  margin-top: 1.5rem;
+}
+
+.payment-actions-buttons .btn {
+  width: 100%;
+  justify-content: center;
+  padding: 1rem;
+  font-size: 1rem;
+}
