+++ frontend/src/components/StatCard/style.css
@@ -1,0 +1,32 @@
+.stat-card {
+  background: var(--card-bg);
+  padding: 2rem;
+  border-radius: 8px;
+  box-shadow: var(--shadow);
+  position: relative;
+  transition: transform 0.1s ease-out;
+  will-change: box-shadow;
+}
+
+.stat-card.highlight {
+  background: linear-gradient(135deg, #fff9e6 0%, var(--card-bg) 100%);
+  border-left: 4px solid var(--primary-color);
+}
+
+.stat-card.highlight .stat-value {
+  color: #f39c12;
+}
+
+.stat-card h3 {
+  font-size: 0.9rem;
+  color: #666;
+  text-transform: uppercase;
+  letter-spacing: 0.5px;
+  margin-bottom: 0.5rem;
+}
+
+.stat-value {
+  font-size: 2rem;
+  font-weight: bold;
+  color: var(--primary-color);
+}
