+++ frontend/src/pages/BillDetails.css
@@ -1,0 +1,198 @@
+/* Bill Details Page Styles */
+.bill-details {
+  max-width: 1000px;
+  margin: 0 auto;
+}
+
+.back-btn {
+  display: inline-flex;
+  align-items: center;
+  gap: 0.5rem;
+  font-size: 0.9rem;
+  margin-bottom: 1rem;
+}
+
+.back-btn:hover {
+  transform: translateX(-2px);
+}
+
+/* Bill Information Card */
+.bill-info-card {
+  background: var(--card-bg);
+  border-radius: 12px;
+  box-shadow: var(--shadow);
+  margin-bottom: 2rem;
+  padding: 2rem;
+  position: relative;
+  transition: transform 0.1s ease-out;
+  will-change: box-shadow;
+}
+
+.bill-info-card.due-today {
+  background: linear-gradient(135deg, #fff9e6 0%, var(--card-bg) 100%);
+  border-left: 4px solid #f39c12;
+}
+
+.bill-info-card.overdue {
+  background: linear-gradient(135deg, #ffe6e6 0%, var(--card-bg) 100%);
+  border-left: 4px solid #e74c3c;
+}
+
+.bill-info-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 2rem;
+  padding-bottom: 1rem;
+  border-bottom: 1px solid var(--border-color);
+}
+
+.bill-info-header h3 {
+  margin: 0;
+  font-size: 2rem;
+  color: var(--text-color);
+}
+
+.status-badge {
+  padding: 0.5rem 1rem;
+  border-radius: 25px;
+  font-size: 0.9rem;
+  font-weight: 600;
+  text-transform: uppercase;
+  letter-spacing: 0.5px;
+}
+
+.status-badge.paid {
+  background: #d4edda;
+  color: #155724;
+}
+
+.status-badge.unpaid {
+  background: #f8d7da;
+  color: #721c24;
+}
+
+.bill-info-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+  gap: 1.5rem;
+}
+
+.info-item {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.info-item label {
+  font-size: 0.85rem;
+  color: #666;
+  text-transform: uppercase;
+  letter-spacing: 0.5px;
+  font-weight: 600;
+}
+
+.info-item div {
+  font-size: 1.1rem;
+  color: var(--text-color);
+  font-weight: 500;
+}
+
+.amount-display {
+  font-size: 1.8rem !important;
+  font-weight: bold !important;
+  color: var(--primary-color) !important;
+}
+
+.overdue-text {
+  color: #e74c3c !important;
+  font-weight: 600 !important;
+}
+
+.due-today-text {
+  color: #f39c12 !important;
+  font-weight: 600 !important;
+}
+
+.notes-item {
+  grid-column: 1 / -1;
+}
+
+.notes-item div {
+  font-style: italic;
+  color: #666;
+  line-height: 1.6;
+}
+
+/* Payments Section */
+.payments-section {
+  background: var(--card-bg);
+  border-radius: 12px;
+  box-shadow: var(--shadow);
+  padding: 2rem;
+}
+
+.section-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 2rem;
+  padding-bottom: 1rem;
+  border-bottom: 1px solid var(--border-color);
+}
+
+.section-header h3 {
+  margin: 0;
+  font-size: 1.5rem;
+  color: var(--text-color);
+}
+
+/* Delete confirmation content for payments */
+.payment-details {
+  background: var(--bg-color);
+  padding: 1rem;
+  border-radius: 6px;
+  margin: 1rem 0;
+  font-family: monospace;
+  font-size: 0.9rem;
+}
+
+/* Responsive Design for Bill Details */
+@media (max-width: 768px) {
+  .bill-info-grid {
+    grid-template-columns: 1fr;
+  }
+
+  .bill-info-header {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 1rem;
+  }
+
+  .section-header {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 1rem;
+  }
+}
+
+/* Small screen optimizations */
+@media (max-width: 480px) {
+  .bill-details {
+    margin: 0 -1rem;
+  }
+
+  .bill-info-card,
+  .payments-section {
+    margin: 0 1rem 2rem 1rem;
+    padding: 1.5rem;
+  }
+
+  .amount-display {
+    font-size: 1.5rem !important;
+  }
+
+  .bill-info-header h3 {
+    font-size: 1.5rem;
+  }
+}
