+++ frontend/src/pages/Bills.css
@@ -1,0 +1,6 @@
+/* Bills Page Styles */
+.bills-list {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+  gap: 1.5rem;
+}
