+++ frontend/src/pages/Categories.css
@@ -1,0 +1,6 @@
+/* Categories Page Styles */
+.categories-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
+  gap: 1rem;
+}
