+++ frontend/src/pages/NotFound.css
@@ -1,0 +1,14 @@
+/* Not Found Page Styles */
+.not-found {
+  text-align: center;
+  padding: 4rem 2rem;
+}
+
+.not-found h2 {
+  margin-bottom: 1rem;
+}
+
+.not-found p {
+  margin-bottom: 2rem;
+  color: #666;
+}
